#50: Refactor to env_manage.py.
