+++ tests/conftest.py
@@ -7,7 +7,7 @@ import shutil
 import pytest
 from requests import HTTPError
 
-from raindrop_todoist_syncer.raindrop import EnvironmentVariablesFileManager
+from raindrop_todoist_syncer.env_manage import EnvironmentVariablesFileManager
 from raindrop_todoist_syncer.rd_credentials import RaindropCredentialsManager
 from raindrop_todoist_syncer.rd_token import RaindropAccessTokenRefresher
 from raindrop_todoist_syncer.rd_class import Raindrop
