+++ tests/unit/test_raindrop_new_access_token_getter.py
@@ -2,9 +2,9 @@ from unittest.mock import patch
 
 import pytest
 
+from raindrop_todoist_syncer.env_manage import EnvironmentVariablesFileManager
 from raindrop_todoist_syncer.rd_credentials import RaindropCredentialsManager
 from raindrop_todoist_syncer.rd_token import RaindropNewAccessTokenGetter
-from raindrop_todoist_syncer.raindrop import EnvironmentVariablesFileManager
 
 
 @pytest.fixture
